Change exp member to uint32 in monster and monstedata. Change saving/loading accordingly

+++ Source/loadsave.cpp
@@ -638,7 +638,7 @@ void LoadMonster(LoadHelper *file, Monster &monster)
 	monster.whoHit = file->NextLE<int8_t>();
 	monster.level = file->NextLE<int8_t>();
 	file->Skip(1); // Alignment
-	monster.exp = file->NextLE<uint16_t>();
+	monster.exp = file->NextLE<uint32_t>();
 
 	if ((monster.flags & MFLAG_GOLEM) != 0) // Don't skip for golems
 		monster.toHit = file->NextLE<uint8_t>();
@@ -1388,7 +1388,7 @@ void SaveMonster(SaveHelper *file, Monster &monster)
 	file->WriteLE<int8_t>(monster.whoHit);
 	file->WriteLE<int8_t>(monster.level);
 	file->Skip(1); // Alignment
-	file->WriteLE<uint16_t>(monster.exp);
+	file->WriteLE<uint32_t>(monster.exp);
 
 	file->WriteLE<uint8_t>(static_cast<uint8_t>(std::min<uint16_t>(monster.toHit, std::numeric_limits<uint8_t>::max()))); // For backwards compatibility
 	file->WriteLE<uint8_t>(monster.minDamage);
